+++ docs/src/markdown/about/changelog.md
@@ -11,6 +11,7 @@ icon: lucide/scroll-text
 -   **NEW**: Tighten restrictions such that `namespaces` and `custom` objects must always be a Mapping, previously lists
     of tuples were also allowed.
 -   **NEW**: Use a singleton for null selectors internally via called `Null` of type `SelectorNull`.
+-   **NEW**: For performance, Soup Sieve will no longer try and coerce bad attribute values to useable strings.
 -   **FIX**: Ensure `custom` is properly passed down from API functions to compilation.
 
 ## 2.9.2